requirements.txt
astroid==2.12.5
attrs==22.1.0
black==22.6.0
cfgv==3.3.1
click==8.1.3
dill==0.3.5.1
distlib==0.3.6
Faker==14.1.0
filelock==3.8.0
flake8==5.0.4
Flask==2.2.2
Flask-Cors==3.0.10
greenlet==1.1.3
identify==2.5.3
importlib-metadata==4.12.0
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
pylint==2.15.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.40
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
virtualenv==20.16.3
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
